+++ crates/codebook/src/queries.rs
@@ -130,6 +130,11 @@ pub static LANGUAGE_SETTINGS: &[LanguageSetting] = &[
             (string_fragment) @string
             (variable_declarator
                 name: (identifier) @identifier)
+            (object
+                (pair
+                key: (property_identifier) @property_name))
+            (catch_clause
+                parameter: (identifier) @identifier)
             (jsx_text) @string
             (shorthand_property_identifier) @identifier
             (function_declaration
@@ -153,6 +158,16 @@ pub static LANGUAGE_SETTINGS: &[LanguageSetting] = &[
             (string_fragment) @string
             (variable_declarator
                 name: (identifier) @identifier)
+            (object
+                (pair
+                key: (property_identifier) @property_name))
+            (interface_declaration
+                name: (type_identifier) @identifier)
+            (interface_body
+                (property_signature
+                    name: (property_identifier) @property_name))
+            (catch_clause
+                parameter: (identifier) @identifier)
             (jsx_text) @string
             (shorthand_property_identifier) @identifier
             (function_declaration